Mobile working too

+++ database/migrations/2025_12_21_141219_create_exercises_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
 
             $table->enum('unit', WorkoutUnits::values());
 
-            $table->unsignedInteger('calories_per_unit')->default(0);
+            $table->float('calories_per_unit')->default(0);
 
             $table->text('note')->nullable();
